Code refactoring
-refactored code to simply the main code
