+++ src/Makefile
@@ -1,1 +1,1 @@
-obj-y += main.o sharedmemory_com.o
+obj-y += main.o sharedmemory_com.o soc_ctrl.o cdcacm_serial.o
